# Conflicts:
#	marketing-check/src/main/resources/prod-istio/scheduler.xml
#	marketing-check/src/main/resources/prod/scheduler.xml
#	marketing-data-monkey/src/main/resources/prod/application-prod.yml
#	marketing-data-monkey/src/main/resources/prod/scheduler.xml
#	marketing-task/src/main/resources/prod/scheduler.xml
#	marketing-xc-general/src/main/resources/prod/scheduler.xml
#	marketing-xc-loop-cycle/src/main/resources/prod/scheduler.xml

+++ marketing-cs/src/main/java/com/br/marketing/mapper/MarketingCustomerConfigMapperBase.java
@@ -2,12 +2,11 @@ package com.br.marketing.mapper;
 
 import com.br.marketing.entity.MarketingCustomerConfig;
 import com.br.marketing.entity.MarketingCustomerConfigExample;
+import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
-import java.util.List;
-
 public interface MarketingCustomerConfigMapperBase {
-    int countByExample(MarketingCustomerConfigExample example);
+    long countByExample(MarketingCustomerConfigExample example);
 
     int deleteByExample(MarketingCustomerConfigExample example);
